rocclr: migrate buildStatus_ from int32_t/CL_BUILD_* to amd::BuildStatus
Change the buildStatus_ field and its accessor in device::Program from
int32_t to the strongly-typed amd::BuildStatus enum class.  Replace all
CL_BUILD_NONE/IN_PROGRESS/ERROR/SUCCESS literals in devprogram.cpp with
the corresponding amd::BuildStatus enumerators.  Update the three
BuildNone guard comparisons in platform/program.cpp to compare directly
against amd::BuildStatus::BuildNone (removing the now-unnecessary
static_cast<int32_t>).  Add an explicit static_cast<cl_build_status> at
the OpenCL API boundary in cl_program.cpp where the enum value is
returned to the caller.

Changes:
- devprogram.hpp: buildStatus_ field and accessor now use amd::BuildStatus
- devprogram.cpp: ~38 CL_BUILD_* assignments/comparisons replaced; buildError_ assignments left as int32_t (separate field)
- platform/program.cpp: 3 BuildNone comparisons simplified to direct enum comparison
- cl_program.cpp: explicit cast at ABI boundary

+++ projects/clr/rocclr/platform/program.cpp
@@ -217,7 +217,7 @@ amd::Status Program::compile(const std::vector<Device*>& devices, size_t numHead
       continue;
     }
     // We only build a Device-Program once
-    if (devProgram->buildStatus() != static_cast<int32_t>(amd::BuildStatus::BuildNone)) {
+    if (devProgram->buildStatus() != amd::BuildStatus::BuildNone) {
       continue;
     }
     if (sourceCode_.empty()) {
@@ -310,7 +310,7 @@ amd::Status Program::link(const std::vector<Device*>& devices, size_t numInputs,
     }
 
     // We only build a Device-Program once
-    if (devProgram->buildStatus() != static_cast<int32_t>(amd::BuildStatus::BuildNone)) {
+    if (devProgram->buildStatus() != amd::BuildStatus::BuildNone) {
       continue;
     }
     amd::Status result =
@@ -457,7 +457,7 @@ amd::Status Program::build(const std::vector<Device*>& devices, const char* opti
     }
 
     // We only build a Device-Program once
-    if (devProgram->buildStatus() != static_cast<int32_t>(amd::BuildStatus::BuildNone)) {
+    if (devProgram->buildStatus() != amd::BuildStatus::BuildNone) {
       continue;
     }
     amd::Status result =
